mental_agent/requirements.txt
requests>=2.31.0
aiohttp>=3.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pymongo>=4.5.0
psycopg2-binary>=2.9.0
PyMySQL>=1.1.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wordcloud>=1.9.0
jieba>=0.42.1
snownlp>=0.12.3
baidu-aip>=4.16.0
openai>=1.0.0
edge-tts>=6.1.0
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0
redis>=5.0.0
celery>=5.3.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
schedule>=1.2.0
apscheduler>=3.10.0
tzlocal>=5.2
pytz>=2023.3
loguru>=0.7.0
colorlog>=6.7.0
tqdm>=4.66.0
rich>=13.6.0
click>=8.1.0
cozepy>=0.5.0